backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
minio==7.2.18
python-multipart==0.0.20
email-validator==2.2.0
celery==5.4.0
redis==5.2.1
sentence-transformers==3.3.1
transformers==4.45.2
Pillow==10.4.0
open-clip-torch==3.3.0
PyPDF2==3.0.1
python-docx==1.0.1
